Fix client bundle: move tutorialCompletionPct to a DB-free module
AnalyticsStatTiles is a client component; value-importing tutorialCompletionPct
from workshopAnalyticsDb pulled the Node DB client (postgres/better-sqlite3)
into the browser bundle and broke the production build (tsc misses this —
ignoreBuildErrors + it's a bundler boundary). Move the pure helper to
lib/workshopCompletion.ts (only an `import type` of the analytics shape) so the
tile and the test import it without the DB chain. Verified with next build.
